requirements.txt
APScheduler==3.6.3
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
cryptography==3.4.4
idna==2.10
pycparser==2.20
python-dotenv==0.15.0
python-telegram-bot==13.2
pytz==2021.1
requests==2.25.1
six==1.15.0
soupsieve==2.2
tornado==6.1
tzlocal==2.1
urllib3==1.26.3
